Add other tests to default e2e test
