+++ e2e-test/default-webpack/test/combine.test.js
@@ -1,24 +1,4 @@
-import { combine } from './index';
-
-const sleep = (ms) => new Promise((r) => setTimeout(r, ms));
-
-describe('Basic test functions', () => {
-	it('should work', () => {
-		expect(1).toEqual(1);
-	});
-
-	it('should handle deep equality', () => {
-		expect({ foo: 1 }).toEqual({ foo: 1 });
-	});
-
-	it('should handle async tests', async () => {
-		let start = Date.now();
-		await sleep(100);
-
-		let now = Date.now();
-		expect(now - start).toBeGreaterThan(50);
-	});
-});
+import { combine } from '../src/index';
 
 describe('combine', () => {
 	it('should concatenate strings', () => {
