+++ e2e-test/default-webpack/test/jest-style.test.js
@@ -1,0 +1,9 @@
+describe('jest-style', () => {
+	describe('not.stringContaining', () => {
+		const expected = 'Hello world!';
+
+		it('matches if the received value does not contain the expected substring', () => {
+			expect('How are you?').toEqual(expect.not.stringContaining(expected));
+		});
+	});
+});
